Passthru .logger= and .logger to ruby-nsq
Here we allow for a passthru of .logger= and .logger to the
wistia/ruby-nsq gem so that we can override the default logger of
Logger.new(nil). This should allow us to setup the nsq gem logger to be
something like Rails.logger or some other logger as is needed.

Changes:
* Add two methods on MessageQueue (.logger= and .logger)
* Support those methods on FakeMessageQueue
* Tests for these methods