call Thread#join only on threads we created
We where using `Thread.list` which could likely contain things we did not
start. This switches to saving the threads we create in an array and
calling `Thread#join` on only the ones we have started.

We could select only non-dead threads from this list by doing
```
listener_threads.select(&:status).map(&:join)
```
but by including threads that have already died, we allow them to
express whatever exception might have been raised back to the main
thread effectly halting the main Thread as they are left unhandled
there.`Thread.list` would have not returned "dead" threads here as we
where using it, so it was possible for a Thread to fail immediately and
have `join` never be called on it making it's failure silent.