Ensure Thread::Queue gets closed
producer.write returns a Thread::Queue which persists until garbage collection, potentially increasing globally open threads to the maximum, which can be problematic.

In applications that persist a singleton producer for most writes, this likely isn't a problem; however, in applications that instantiate individual 'one-off' producers (eg, for a rake task), this can lead to resource starvation.

Note that this is effectively opt-in: downstream apps would need to switch from `producer.write` to `messenger.deliver` to get the auto-closing Thread::Queue.