Setup a possible path to daemonize listeners
This is a rip off of Sidekiqs cli made to work with our existing
listeners. It still creates one listener per setup topic, but does so in
a bit more of a managed way. There is quite a bit to be done in the way
of cleanup and ... I have yet to test anything because I wanted feedback
on the direction before spending the time.