only raise if consumer is empty
terminate should allow a consumer to attempt to finish the work that it
is doing. We want it to finish the message it is on and then on the next
pass of the loop, see that @done is set and finish up. However, if
there are no messages, consumer.pop will be blocking and will never find
out that @done is set. We will hit the stop deadline and end up calling
kill. Here we are adding a check to see if the consumer is empty, and
only if it is do we raise to get it out from being stuck on a pop.
Otherwise, we allow it to finish the work it is on and findout that
@done is set ...