refactor: use sidekiq-style testing
Messages are now feed directly into the manager's thread pool and will
not aggregate on the individual listeners.  Using Sidekiq-style testing
enables developers to control the behavior without referencing the
listener instances or expecting #pop to be available.

This module monkey patches nsq interface classes and short circuits to
super when testing is disabled.  The risk of these patches is mitigated
by the live testing that is performed by the interface classes.  If it
works with the patches, it will work without them.